requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2
python-dotenv>=0.19.0

# Security and authentication
python-jose>=3.3.0
#passlib>=1.7.4
#bcrypt>=3.2.0
bcrypt==4.0.1
passlib[bcrypt]

# OCR libraries
pytesseract>=0.3.8
Pillow>=8.2.0
#pdf2image>=1.16.0

# Streamlit demo app
streamlit>=1.0.0
requests>=2.26.0

# Testing
pytest>=6.2.5
httpx>=0.19.0
pytest-cov>=2.12.1